feature-03 무한스크롤 최적화 및 Event 탭 컨텐츠 추가 (#3)
* 코드 정리

* 무한스크롤 컨텐츠 스켈레톤 추가

* react query추가

* 이미지 webp 추가

* 테스트 위한 delay 초 수정

* event 컨텐츠 추가

* 코드 정리
