+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.70.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "sass": "^1.86.0",
