+++ public/json/banner.json
@@ -6,7 +6,7 @@
       "startDate": "2025-03-20",
       "endDate": "2025-04-20",
       "type": "vote",
-      "thumbnail": "/images/event1.jpg",
+      "thumbnail": "https://picsum.photos/360/130?id=1",
       "url": "https://www.hanteochart.com/",
       "status": "ongoing"
     },
@@ -16,7 +16,7 @@
       "startDate": "2025-03-30",
       "endDate": "2025-04-30",
       "type": "fan-club",
-      "thumbnail": "/images/event2.jpg",
+      "thumbnail": "https://picsum.photos/360/130?id=2",
       "url": "https://www.hanteochart.com/",
       "status": "ongoing"
     },
@@ -26,7 +26,7 @@
       "startDate": "2025-04-25",
       "endDate": "2025-04-25",
       "type": "showcase",
-      "thumbnail": "/images/event3.jpg",
+      "thumbnail": "https://picsum.photos/360/130?id=3",
       "url": "https://www.hanteochart.com/",
       "status": "upcoming"
     },
@@ -36,7 +36,7 @@
       "startDate": "2025-02-15",
       "endDate": "2025-03-15",
       "type": "fan-meeting",
-      "thumbnail": "/images/event4.jpg",
+      "thumbnail": "https://picsum.photos/360/130?id=4",
       "url": "https://www.hanteochart.com/",
       "status": "ended"
     }
