+++ src/api/chart.ts
@@ -1,5 +1,17 @@
-export const getChart = async () => {
-  const response = await fetch('/json/chart.json');
-  const data = await response.json();
-  return data;
+import { Chart } from '@type';
+
+export const getChart = async (page: number, pageSize: number) => {
+  return new Promise<{ chart: Chart[] }>((resolve) => {
+    setTimeout(async () => {
+      const response = await fetch('/json/chart.json');
+      const data = await response.json();
+
+      const startIndex = (page - 1) * pageSize;
+      const endIndex = startIndex + pageSize;
+
+      const chart = data.chart.slice(startIndex, endIndex);
+
+      resolve({ chart });
+    }, 1000); // 테스트를 위한 딜레이 추가
+  });
 };
