+++ src/main.tsx
@@ -5,11 +5,16 @@ import '@/styles/index.scss';
 import App from './App.tsx';
 import { colors } from './styles/color.ts';
 import { fonts } from './styles/font.ts';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+const queryClient = new QueryClient();
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <ThemeProvider theme={{ colors, fonts }}>
-      <App />
-    </ThemeProvider>
+    <QueryClientProvider client={queryClient}>
+      <ThemeProvider theme={{ colors, fonts }}>
+        <App />
+      </ThemeProvider>
+    </QueryClientProvider>
   </StrictMode>,
 );
